+++ review-service/src/main/java/com/prs/rs/aop/MemberValidateAspect.java
@@ -1,8 +1,8 @@
 package com.prs.rs.aop;
 
-import com.prs.rs.annotation.ValidatePlatform;
-import com.prs.rs.client.PlatformServiceClient;
-import com.prs.rs.dto.response.PlatformInfoDto;
+import com.prs.rs.annotation.ValidateMember;
+import com.prs.rs.client.MemberServiceClient;
+import com.prs.rs.dto.response.MemberInfoDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -19,11 +19,11 @@ import java.lang.reflect.Method;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ValidatePlatformAspect {
+public class MemberValidateAspect {
 
-    private final PlatformServiceClient platformServiceClient;
+    private final MemberServiceClient memberServiceClient;
 
-    @Around("execution(* *(.., @com.prs.rs.annotation.ValidatePlatform (*), ..))")
+    @Around("execution(* *(.., @com.prs.rs.annotation.ValidateMember (*), ..))")
     public Object retry(ProceedingJoinPoint joinPoint) throws Throwable{
 
 
@@ -37,31 +37,15 @@ public class ValidatePlatformAspect {
         // 파라미터 배열을 탐색하여 @ValidatePlatform 어노테이션이 있는 PlatformInfoDto 타입의 파라미터를 찾아서 검증
         Loop : for (int i = 0; i < parameterAnnotations.length; i++) {
             for (Annotation annotation : parameterAnnotations[i]) {
-                if (annotation instanceof ValidatePlatform && parameters[i] instanceof Long platformId) {
+                if (annotation instanceof ValidateMember && parameters[i] instanceof MemberInfoDto) {
 
-                    PlatformInfoDto platformInfo = platformServiceClient.getPlatformInfo(platformId);
-
-                    Integer platformParameterIdx = findParameterIdx(parameters);
-
-                    parameters[platformParameterIdx] = platformInfo;
+                    MemberInfoDto memberInfo = memberServiceClient.getMemberInfo();
+                    parameters[i] = memberInfo;
 
                     break Loop;
                 }
             }
         }
         return joinPoint.proceed(parameters);
-    }
-
-
-    /*
-     * PlatformInfoDto 타입이 있는 배열 인덱스 반환
-     */
-    private Integer findParameterIdx(Object[] parameters) {
-        for (int i = 0; i < parameters.length; i++) {
-            if(parameters[i] instanceof PlatformInfoDto) {
-                return i;
-            }
-        }
-        return -1;
     }
 }
