+++ review-service/src/main/java/com/prs/rs/aop/PlatformValidateAspect.java
@@ -1,10 +1,8 @@
-package com.prs.ps.aop;
+package com.prs.rs.aop;
 
-
-import com.prs.ps.annotation.ValidatePlatform;
-import com.prs.ps.domain.Platform;
-import com.prs.ps.exception.PlatformNotFoundException;
-import com.prs.ps.repository.PlatformRepository;
+import com.prs.rs.annotation.ValidatePlatform;
+import com.prs.rs.client.PlatformServiceClient;
+import com.prs.rs.dto.response.PlatformInfoDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -12,37 +10,40 @@ import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.reflect.MethodSignature;
 import org.springframework.stereotype.Component;
+
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
-import java.util.Optional;
 
 
 @Aspect
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ValidatePlatformAspect {
+public class PlatformValidateAspect {
 
-    private final PlatformRepository platformRepository;
+    private final PlatformServiceClient platformServiceClient;
 
-    @Around("execution(* *(.., @com.prs.ps.annotation.ValidatePlatform (*), ..))")
-    public Object validate(ProceedingJoinPoint joinPoint) throws Throwable {
+    @Around("execution(* *(.., @com.prs.rs.annotation.ValidatePlatform (*), ..))")
+    public Object retry(ProceedingJoinPoint joinPoint) throws Throwable{
+
+
         MethodSignature methodSignature = (MethodSignature) joinPoint.getSignature();
         Method method = methodSignature.getMethod();
-        Annotation[][] parameterAnnotations = method.getParameterAnnotations(); // 각 파라미터에 달린 어노테이션 가져오기
+        Annotation[][] parameterAnnotations = method.getParameterAnnotations(); // 파라미터에 달린 어노테이션 가져오기
+
+        Object[] parameters = joinPoint.getArgs(); // 파라미터
 
 
-        Object[] parameters = joinPoint.getArgs();
-
-
-        // 파라미터 배열을 탐색하여 @ValidatePlatform 어노테이션이 있는 Long 타입의 파라미터를 찾아서 플랫폼 검증
+        // 파라미터 배열을 탐색하여 @ValidatePlatform 어노테이션이 있는 PlatformInfoDto 타입의 파라미터를 찾아서 검증
         Loop : for (int i = 0; i < parameterAnnotations.length; i++) {
             for (Annotation annotation : parameterAnnotations[i]) {
                 if (annotation instanceof ValidatePlatform && parameters[i] instanceof Long platformId) {
 
-                    Optional<Platform> platform = platformRepository.findById(platformId);
+                    PlatformInfoDto platformInfo = platformServiceClient.getPlatformInfo(platformId);
+
                     Integer platformParameterIdx = findParameterIdx(parameters);
-                    parameters[platformParameterIdx] = platform.orElseThrow(PlatformNotFoundException::new);
+
+                    parameters[platformParameterIdx] = platformInfo;
 
                     break Loop;
                 }
@@ -51,12 +52,13 @@ public class ValidatePlatformAspect {
         return joinPoint.proceed(parameters);
     }
 
+
     /*
-     * Platform 타입이 있는 배열 인덱스 반환
+     * PlatformInfoDto 타입이 있는 배열 인덱스 반환
      */
     private Integer findParameterIdx(Object[] parameters) {
         for (int i = 0; i < parameters.length; i++) {
-            if(parameters[i] instanceof Platform) {
+            if(parameters[i] instanceof PlatformInfoDto) {
                 return i;
             }
         }
